requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
backoff==2.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==7.1.2
click-params==0.3.0
cohere==4.34
cryptography==41.0.5
decorator==5.1.1
fastavro==1.8.2
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
multidict==6.0.4
pycparser==2.21
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
tqdm==4.66.1
urllib3==2.0.7
validators==0.18.2
weaviate-cli==2.2.0
weaviate-client==3.23.2
yarl==1.9.2
zipp==3.17.0
